+++ crates/nostr/CHANGELOG.md
@@ -31,6 +31,7 @@
 
 ### Added
 
+- Impl `TryFrom<i64>` for `Timestamp`
 - Add `RelayUrlScheme` enum and `RelayUrl::scheme` method (https://github.com/rust-nostr/nostr/pull/1127) 
 
 ## v0.44.1 - 2025/11/09